+++ src/boxrec.class.ts
@@ -6,7 +6,9 @@ const BoxrecPageProfile = require("./boxrec-pages/boxrec.page.profile");
 
 export class Boxrec {
 
-    private cookie: any = null; // is a string but `getCookieString` returns void?
+    private _cookie: any = null; // is a string but `getCookieString` returns void?
+    private _hasLoggedIn: boolean = false;
+    private _cookieJar: any;
 
     async login(username: string, password: string): Promise<void> {
 
@@ -28,6 +30,8 @@ export class Boxrec {
         cookieJar.setCookie(cookie, "http://boxrec.com", () => {
         });
 
+        this._cookieJar = cookieJar;
+
         const options = {
             uri: "http://boxrec.com/en/login", // boxrec does not support HTTPS
             followAllRedirects: true, // 302 redirect occurs
@@ -39,7 +43,7 @@ export class Boxrec {
                 "_target_path": "http://boxrec.com", // not required
                 "login[go]": "", // not required
             },
-            jar: cookieJar,
+            jar: this._cookieJar,
         };
 
         await rp.post(options)
@@ -56,16 +60,17 @@ export class Boxrec {
                 }
 
                 if (data.statusCode !== 200 || errorMessage !== "") {
-                    throw new Error(errorMessage)
+                    throw new Error(errorMessage);
                 }
             });
 
         // it argues that the return value is void
-        this.cookie = cookieJar.getCookieString("http://boxrec.com", () => {
+        this._cookie = cookieJar.getCookieString("http://boxrec.com", () => {
             // the callback doesn't work but it works if you assign as a variable?
         });
 
-        if (this.cookie && this.cookie.includes("PHPSESSID") && this.cookie.includes("REMEMBERME")) {
+        if (this._cookie && this._cookie.includes("PHPSESSID") && this._cookie.includes("REMEMBERME")) {
+            this._hasLoggedIn = true;
             return; // success
         } else {
             throw new Error("Cookie did not have PHPSESSID and REMEMBERME");
@@ -73,8 +78,10 @@ export class Boxrec {
     }
 
     async getBoxerById(boxerId: number) {
+        this.checkIfLoggedIn();
         const boxrecPageBody = await rp.get({
-            uri: `http://boxrec.com/en/boxer/${boxerId}`
+            uri: `http://boxrec.com/en/boxer/${boxerId}`,
+            jar: this._cookieJar,
         });
 
         return new BoxrecPageProfile(boxrecPageBody);
@@ -86,6 +93,14 @@ export class Boxrec {
             uri: "http://boxrec.com",
             resolveWithFullResponse: true,
         }).then((data: RequestResponse) => data.headers["set-cookie"]);
+    }
+
+    private checkIfLoggedIn(): boolean {
+        if (this._hasLoggedIn) {
+            return true;
+        } else {
+            throw new Error("This package requires logging into Boxrec to work properly.  Please use the `login` method before any other calls");
+        }
     }
 
 }