+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.3.0 (2018-03-27)
+
+### Added
+
+- Profile `bouts` will return an array of basic information about the bout including weight, result, location, judges, referees
+
+### Changed
+
+- Getting the profile `bouts` used to give the number of bouts the fighter has completed.  `bouts` is now an array of basic bout information.  `numberOfBouts` will now give you the amount of bouts that the boxer has had
+
 ## 0.2.1 (2018-03-22)
 
 ### Added